Update templates and scripts for Ubuntu 14.04/15.10
Update JSON templates for Ubuntu 14.04.3 and 14.04.4. Update JSON template for Ubuntu 15.10. Break post-configuration scripts into multiple files for better modularity moving forward.

+++ scripts/vagrant-config.sh
@@ -1,18 +1,17 @@
 #!/bin/sh -x
 
-# Install Vagrant insecure public key
+# Create private .ssh directory for 'vagrant' user
 mkdir -p /home/vagrant/.ssh
-chown vagrant:vagrant /home/vagrant/.ssh
-chmod 700 /home/vagrant/.ssh
-cd /home/vagrant/.ssh
-mv /home/vagrant/authorized_keys /home/vagrant/.ssh/authorized_keys
-chmod 600 /home/vagrant/.ssh/authorized_keys
-chown -R vagrant:vagrant /home/vagrant/.ssh
+chmod 0700 /home/vagrant/.ssh
 
-# Create /vagrant directory
+# Install Vagrant insecure public key
+if [ -f /home/vagrant/authorized_keys ]; then
+    mv /home/vagrant/authorized_keys /home/vagrant/.ssh/authorized_keys
+    chmod 0600 /home/vagrant/.ssh/authorized_keys
+fi
+
+# Set ownership for private .ssh directory and contents
+sudo chown -R vagrant:vagrant /home/vagrant.ssh
+
+# Create /vagrant directory (for shared folders)
 mkdir -p /vagrant
-
-# Install sudoers file for Vagrant
-mv /home/vagrant/vagrant /etc/sudoers.d/vagrant
-chown root:root /etc/sudoers.d/vagrant
-chmod 0440 /etc/sudoers.d/vagrant
